There are three changes here: 1. Change data type of HKL values to float16 from float64. This reduces memory consumption by a factor of 4. Might be more efficient to do this in the c routines maybe? 2. Avoid obtaining minimum and maximumv values of HKL if not needed in the grid3d routine. This can be the slowest part. 3. Discard pixels with NAN values in the C routine. This is important to avoid dealing with masked arrays.

+++ skbeam/core/recip.py
@@ -177,7 +177,7 @@ def process_to_q(setting_angles, detector_size, pixel_size,
     logger.info("Processing time for {0} {1} x {2} images took {3} seconds."
                 "".format(setting_angles.shape[0], detector_size[0],
                           detector_size[1], (t2 - t1)))
-    return hkl
+    return hkl.astype(np.float16)
 
 # Assign frame_mode as an attribute to the process_to_q function so that the
 # autowrapping knows what the valid options are
